WEB-7 Formatters are moved to the their own file
